Reduce dependencies and fix deadlocks
Implemented an Antimony specific which crate with caching
Implemented a simple xdg-open substitute to reduce sandbox size
Cleaned up antimony-dumper and replaced our deadlock-prone binary parser. It now terminates after syscalls that indicate a waiting state, such as wait or poll, but its not perfect, and is barred to only work on shell scripts.
Spawn now has an abs constructor, for cases where we provide the absolute path and don't want to waste time resolving it.
Notify now logs the current thread.
The spawn::Handle now supports waiting on timeout using a SIGARLM, terminating the process without consuming the handle, and a more intelligent alive() check that takes ECHILD into account. The handle should now hang much less often.
Move back to parking lot, as deadlock_detection is useful, and the performance is negligble
Remove StreamMode::Discard, as it was causing strange behavior
Fix concurrency issues in the binary and library parser.

+++ src/setup/files.rs
@@ -26,8 +26,10 @@ fn get_x(file: &str, handle: &Spawner) -> Result<()> {
 
 pub fn add_file(handle: &Spawner, file: &str, contents: String, op: FileMode) -> Result<()> {
     let path = direct_path(file);
-    if !path.exists() {
-        fs::create_dir_all(path.parent().unwrap())?;
+    if !path.exists()
+        && let Some(parent) = path.parent()
+    {
+        fs::create_dir_all(parent)?;
         let contents = resolve_env(Cow::Borrowed(&contents));
         fs::write(&path, contents.as_ref())?;
     }
